+++ src/NeuralNet/Optimizers/Momentum.php
@@ -51,14 +51,14 @@ class Momentum implements Optimizer, Adaptive
      * @param float $decay
      * @throws \InvalidArgumentException
      */
-    public function __construct(float $rate = 0.001, float $decay = 0.9)
+    public function __construct(float $rate = 0.001, float $decay = 0.1)
     {
         if ($rate <= 0.) {
             throw new InvalidArgumentException('Learning rate must be'
                 . " greater than 0, $rate given.");
         }
 
-        if ($decay < 0. or $decay > 1.) {
+        if ($decay <= 0. or $decay >= 1.) {
             throw new InvalidArgumentException('Decay must be between'
                 . " 0 and 1, $decay given.");
         }
@@ -88,7 +88,7 @@ class Momentum implements Optimizer, Adaptive
         $velocity = $this->cache[$param->id()];
 
         $velocity = $gradient->multiply($this->rate)
-            ->add($velocity->multiply($this->decay));
+            ->add($velocity->multiply(1. - $this->decay));
 
         $this->cache[$param->id()] = $velocity;